requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
flake8==3.9.2
idna==3.2
mccabe==0.6.1
multidict==5.1.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
typing-extensions==3.10.0.0
yarl==1.6.3
